fixup depcache
