+++ src/features/import-maps.js
@@ -1,8 +1,8 @@
 /*
  * SystemJS browser attachments for script and import map processing
  */
-import { baseUrl, resolveAndComposeImportMap, hasDocument, resolveUrl, IMPORT_MAP } from '../common.js';
-import { systemJSPrototype, getOrCreateLoad } from '../system-core.js';
+import { baseUrl, resolveAndComposeImportMap, hasDocument, resolveUrl } from '../common.js';
+import { systemJSPrototype } from '../system-core.js';
 import { errMsg } from '../err-msg.js';
 
 var importMapPromise = Promise.resolve();
@@ -23,16 +23,6 @@ if (hasDocument) {
   processScripts();
   window.addEventListener('DOMContentLoaded', processScripts);
 }
-
-var systemInstantiate = systemJSPrototype.instantiate;
-systemJSPrototype.instantiate = function (url, firstParentUrl) {
-  var preloads = (!process.env.SYSTEM_BROWSER && this[IMPORT_MAP] || importMap).depcache[url];
-  if (preloads) {
-    for (var i = 0; i < preloads.length; i++)
-      getOrCreateLoad(this, this.resolve(preloads[i], url), url);
-  }
-  return systemInstantiate.call(this, url, firstParentUrl);
-};
 
 function processScripts () {
   [].forEach.call(document.querySelectorAll('script'), function (script) {
