+++ src/s.js
@@ -1,4 +1,5 @@
 import './features/script-load.js';
 import './features/resolve.js';
 import './features/import-maps.js';
+import './features/depcache.js';
 import './features/worker-load.js';
