+++ src/system.js
@@ -1,6 +1,7 @@
 import './features/script-load.js';
 import './features/resolve.js';
 import './features/import-maps.js';
+import './features/depcache.js';
 import './features/worker-load.js';
 import './extras/global.js';
 import './extras/module-types.js';
